+++ Engine/Source/Core/Layer/ImGuiLayer.h
@@ -4,13 +4,15 @@
 
 namespace Core
 {
-    class ImGuiLayer
+    class CE_API ImGuiLayer
     {
     public:
         ImGuiLayer(){};
         ~ImGuiLayer(){};
 
+        
         static void Init();
+        static void SetColorsToDefaultStyle();
         static void Shutdown();
         static void BeginFrame();
         static void EndFrame();