+++ Engine/Source/Renderer/Object/Mesh.h
@@ -2,6 +2,7 @@
 
 #include "Base.h"
 #include "Renderer/Material/Material.h"
+#include "Renderer/Geometry/Geometry.h"
 #include "Renderer/Material/MaterialSystem.h"
 
 namespace Core
@@ -10,9 +11,13 @@ namespace Core
     {
     private:
         class VertexArray *array;
+
         Material *material;
+        Geometry *geometry;
 
         void ReleaseMaterial();
+        void ReleaseGeometry();
+        void SetupVertexArray();
 
     public:
         Mesh();
@@ -27,10 +32,15 @@ namespace Core
 
         /// @brief Will set material to the config, along side releasing the last.
         /// @param config The configuration.
-        void SetMaterial(Material::Configuration* config);
+        void SetMaterial(Material::Configuration *config);
+
+        /// @brief Will set the geometry to the new one. Use "new GEOMETRY_TYPE(ARGS...)" for best results.
+        /// @param _geometry An instance of the desired geometry 
+        void SetGeometry(Geometry* _geometry);
 
 
         inline Material *GetMaterial() { return material; };
+        inline Geometry *GetGeometry() { return geometry; };
 
         void MakeMaterialUnique();
         void MakeMaterialDefault();